deprecate: SolvingFactorizationExecutorをLUTypeSolver.Executorに代替させ,削除予定とする.
これは, SolvingFactorizationExecutorが持つ型パラメータの境界が不適切であるため, インターフェース全体の置き換えを促すdeprecateである.
