+++ src/matsu/num/matrix/core/nlsf/Cholesky.java
@@ -5,7 +5,7 @@
  * http://opensource.org/licenses/mit-license.php
  */
 /*
- * 2024.11.23
+ * 2025.6.26
  */
 package matsu.num.matrix.core.nlsf;
 
@@ -143,7 +143,7 @@ public final class Cholesky
      * 
      */
     public static final class Executor
-            extends SkeletalSolvingFactorizationExecutor<
+            extends SkeletalLUTypeSolver.Executor<
                     EntryReadableMatrix, Cholesky> {
 
         private static final Executor INSTANCE = new Executor();
