+++ src/matsu/num/matrix/core/nlsf/ModifiedCholeskyPivoting.java
@@ -5,7 +5,7 @@
  * http://opensource.org/licenses/mit-license.php
  */
 /*
- * 2024.11.23
+ * 2025.6.26
  */
 package matsu.num.matrix.core.nlsf;
 
@@ -147,7 +147,7 @@ public final class ModifiedCholeskyPivoting
      * 
      */
     public static final class Executor
-            extends SkeletalSolvingFactorizationExecutor<
+            extends SkeletalLUTypeSolver.Executor<
                     EntryReadableMatrix, ModifiedCholeskyPivoting> {
 
         private static final Executor INSTANCE =
